Add files via upload
This is the single Project file that has all the classes  of the project. Run the Main (Project) object of this project to get the enriched trip. Change the file source path for Routes,trips and calendars and choose the path where the output file needs to be stored. The input files will be in STM_GTFS repository.